refactor: move fullscreen functionality to `fullscreen/lib.rs`
