+++ components/script/dom/fullscreen/mod.rs
@@ -2,6 +2,4 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-pub(crate) mod document;
-pub(crate) mod element;
-pub(crate) use self::element::*;
+pub(crate) mod lib;
